Format the output and execute from correct directory

+++ programming_paradigm/bank_account.py
@@ -1,16 +1,16 @@
 class BankAccount:
   def __init__(self, acc_bal=0):
-    acc_bal = float(acc_bal)
+    acc_bal = int(acc_bal)
     self.account_balance = acc_bal
   
   def deposit(self, amount):
-    self.account_balance += amount
+    self.account_balance += int(amount)
   
   def withdraw(self, amount):
-    if self.account_balance < amount:
+    if self.account_balance < int(amount):
       return False
     else:
-      self.account_balance -= amount
+      self.account_balance -= int(amount)
       return True
   
   def display_balance(self):
